+++ frontend/app/navigation/RootNavigator.tsx
@@ -7,7 +7,7 @@ import { TabNavigator } from "./TabNavigator";
 import { Header } from "../components/Header";
 import { InAppNotification } from "../components/InAppNotification";
 import { TutorialOverlay } from "../components/TutorialOverlay";
-import { LoadingScreen } from "../components/LoadingScreen";
+import { LoadingScreen } from "../pages/LoadingScreen";
 
 import { useFocEngine } from "../context/FocEngineConnector";
 import { useTutorial } from "../stores/useTutorialStore";
@@ -30,13 +30,13 @@ export const RootNavigator = memo(() => {
       {isAuthenticated ? (
         // Temporarily showing LoadingScreen for testing
         <LoadingScreen />
+      ) : (
         // <>
         //   {isTutorialActive && <TutorialOverlay />}
         //   <Header />
         //   <InAppNotification />
         //   <TabNavigator />
         // </>
-      ) : (
         <Stack.Navigator screenOptions={screenOptions}>
           <Stack.Screen name="Login" component={LoginPage} />
         </Stack.Navigator>
